hw 1.22 refactoring
