+++ src/cli/commands/import.command.ts
@@ -3,6 +3,8 @@ import {fileRead} from '#src/utils/file-read.js';
 import {tsvOffersParse} from '#src/utils/tsv-offers-parse.js';
 
 export class ImportCommand extends BaseCommand {
+  private readonly _name = '--import';
+
   public async execute(...parameters: string[]): Promise<void> {
     await this.parseFileList(parameters);
   }
@@ -16,6 +18,6 @@ export class ImportCommand extends BaseCommand {
   }
 
   get name(): string {
-    return '--import';
+    return this._name;
   }
 }
