+++ src/cli/commands/version.command.ts
@@ -5,13 +5,10 @@ import chalk from 'chalk';
 
 export class VersionCommand extends BaseCommand {
   private readonly filePath = './package.json';
+  private readonly _name = '--version';
 
   public async execute(..._parameters: string[]): Promise<void> {
     console.info(chalk.blue(await this.readVersion()));
-  }
-
-  get name(): string {
-    return '--version';
   }
 
   private isJsonConfig(obj: unknown): obj is {version?: string} {
@@ -22,10 +19,13 @@ export class VersionCommand extends BaseCommand {
     const rawFileContent = await fileRead(this.filePath);
     const fileConfig = jsonParse(rawFileContent);
     if (!this.isJsonConfig(fileConfig) || typeof fileConfig.version !== 'string') {
-      console.error(`'version' field is missing or not a string in the file at ${this.filePath}`);
-      throw new Error('Failed to parse json content.');
+      throw new Error(`'version' field is missing or not a string in the file at ${this.filePath}`);
     }
     return fileConfig.version;
+  }
+
+  get name(): string {
+    return this._name;
   }
 }
