+++ src/types/user.type.ts
@@ -1,9 +1,12 @@
-export type UserType = 'обычный' | 'pro';
+export enum UserType {
+  basic = 'обычный',
+  pro = 'pro'
+}
 
 export interface User {
   name: string,
   email: string,
-  avatarUrl: string,
+  avatarUrl?: string,
   password: string,
   type: UserType
 }
