--- src/utils/json-parse.ts
@@ -1,10 +1,0 @@
-function jsonParse(jsonString: string): string {
-  try {
-    return JSON.parse(jsonString);
-  } catch (error) {
-    console.error(`Error parsing JSON: ${error}`);
-    return '';
-  }
-}
-
-export {jsonParse};
